refactor(core): satellites merge back into their parents (KEN-547) [no-changelog]
Eleven one-importer splits fold back into the file they were cut from.
Each pair reads as one concept, so the merged rows rise under
RATCHET_RAISE=1 (size-ratchet: merging fragments) and the emptied files
go in the same diff. No behavior change.

- engine/adopt/{capture,declare,position}.rs -> adopt.rs
- quality/author/{read,budget}.rs -> author.rs
- engine/gate/settle.rs -> gate.rs
- engine/desired_agent/project.rs -> desired_agent.rs
- engine/ops/add/{subsume,declare}.rs -> add.rs (pick and place stay)
- remote/store/adopt.rs -> store.rs
- check_catalog/safety.rs -> check_catalog.rs
- engine/desired_skill/authored.rs -> desired_skill.rs
- engine/ops/persist.rs -> ops.rs
- engine/deps/graph.rs -> deps.rs
- scan/entry.rs -> scan/mod.rs

Items that were pub(super) for the parent alone are private now; the
one cross-file reference (adopt_shared's MAX_CAPTURE_FILES) follows the
constant. The doc block that sat on `mod declare;` in add.rs describes
pick::ensure_source and moves there.
