+++ crates/core/src/quality/author/tests.rs
@@ -1,4 +1,3 @@
-use super::read::{honest, is_timestamp};
 use super::*;
 use crate::quality::Severity;
 use crate::quality::reviews::Dismissal;
