+++ tools/size-ratchet-baseline.tsv
@@ -1,5 +1,12 @@
 .github/workflows/review-gate-writer.yml	652
+crates/core/src/check_catalog.rs	408
+crates/core/src/engine/adopt.rs	539
+crates/core/src/engine/desired_agent.rs	525
+crates/core/src/engine/gate.rs	430
+crates/core/src/engine/ops/add.rs	549
 crates/core/src/error.rs	429
+crates/core/src/quality/author.rs	458
+crates/core/src/remote/store.rs	420
 docs/ARCHITECTURE.md	911
 hooks/block-repo-copy.sh	511
 pi-extensions/pi-agents-tmux/extensions/subagent/browser.ts	650
